+++ backend/spec/features/admin/configuration/store_spec.rb
@@ -20,7 +20,6 @@ describe "Store", type: :feature, js: true do
 
   context "visiting general store settings" do
     it "should have the right content" do
-      expect(page).to have_content("SettingsStore")
       expect(page).to have_field("store_name", with: "Test Store")
       expect(page).to have_field("store_url", with: "test.example.org")
       expect(page).to have_field("store_mail_from_address", with: "test@example.org")
@@ -33,7 +32,7 @@ describe "Store", type: :feature, js: true do
       fill_in "store_mail_from_address", with: "spree@example.org"
       click_button "Update"
 
-      assert_successful_update_message(:general_settings)
+      expect(page).to have_content "successfully updated"
       expect(page).to have_field("store_name", with: "Spree Demo Site99")
       expect(page).to have_field("store_mail_from_address", with: "spree@example.org")
     end
@@ -44,18 +43,18 @@ describe "Store", type: :feature, js: true do
       select "Germany", from: "Tax Country for Empty Carts"
       click_button "Update"
 
-      assert_successful_update_message(:general_settings)
+      expect(page).to have_content("has been successfully updated")
       expect(page).to have_select("Tax Country for Empty Carts", selected: "Germany")
     end
   end
 
   context "update fails" do
     it "should display the error" do
-      fill_in "Site Name", with: ""
+      fill_in "Site Name", with: " "
       click_button "Update"
 
       expect(page).to have_content("can't be blank")
-      expect(page).to have_field("Site Name", with: "")
+      expect(page).to have_field("Site Name", with: " ")
     end
   end
 end